refactor #107: move NotificationServiceMock
