+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/notification/service/NotificationService.java
@@ -1,4 +1,4 @@
-package swm.hkcc.LGTM.app.global.notification.service;
+package swm.hkcc.LGTM.app.modules.notification.service;
 
 import java.util.Map;
