+++ API-Server/src/test/java/swm/hkcc/LGTM/app/global/api/IntroControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import swm.hkcc.LGTM.app.global.notification.service.NotificationServiceImpl;
+import swm.hkcc.LGTM.app.modules.notification.service.NotificationServiceImpl;
 
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
 import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
