+++ API-Server/src/test/java/swm/hkcc/LGTM/app/global/api/IntroPushControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import swm.hkcc.LGTM.app.global.notification.service.NotificationServiceImpl;
+import swm.hkcc.LGTM.app.modules.notification.service.NotificationServiceImpl;
 import swm.hkcc.LGTM.utils.CustomMDGenerator;
 
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
