+++ API-Server/src/test/java/swm/hkcc/LGTM/app/global/notification/service/NotificationServiceMock.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
+import swm.hkcc.LGTM.app.modules.notification.service.NotificationService;
 
 import java.util.Map;
